More Aura 1.1 stuff
Introduces a ton of missing stuff and also adds CTF definitions for some of the maps. Positioning of the flags is the same as they were in the original Halo games. Also replaces the AG announcer with the iconic Jeff Steitzer. Some lines may or may not change in the future.

Create Damned_Event.ctf

Update .gitignore